requirements.txt
pydantic>=2.0.0; python_version < "3.13.0"
pyyaml; python_version < "3.13.0"
Pillow>=9.0.0; python_version < "3.13.0"
numpy>=1.23.0; python_version < "3.13.0"
numba>=0.50.0; python_version < "3.13.0"
opencv-python; python_version < "3.13.0"
coremltools>=6.0.0; python_version < "3.13.0"
onnxruntime>=1.15.0; python_version < "3.13.0"
